First cut at replacing make with bazel.
Tried to untangle a lot of dependences.
But everything in libsim is still pretty much a mess.
I wouldn't count on incremental builds wrt to headers just yet until we fix this.

Quite a few things still missing:
- test rules
- O3 builds / 32-bit builds; in fact, a top-level script to set up flags
- debug flags
- ztrace
- a proper libconfuse dependence

BUG=XIOSIM-15
TESTED= cd pintool; bazel build --cpu=k8 -c opt --dynamic_mode off --copt="-Wno-unused-function" ... ; ./run.sh

+++ mcpat/slave.cc
@@ -99,7 +99,7 @@ void mcpat_compute_energy(bool print_power, double * cores_rtp, double * uncore_
         proc->cores[i]->l2cache->cachep.executionTime = proc->XML->sys.core[i].total_cycles/(proc->cores[i]->coredynp.clockRate * 1e6);
    }
 
-   for (i=0; i<proc->l2array.size();i++)
+   for (i=0; i<(int)proc->l2array.size();i++)
       proc->l2array[i]->cachep.executionTime = proc->XML->sys.total_cycles/(proc->XML->sys.L2[i].clockrate*1e6);
    for (i=0; i<proc->numL1Dir;i++)
       proc->l1dirarray[i]->cachep.executionTime = proc->XML->sys.total_cycles/(proc->XML->sys.L2[i].clockrate*1e6);
